Reduce mobile CMT overload with 5/1/1 topology and batched provision
Mobile CMT now provisions full suite (5) then smoke (1 each) under a
global lock, refuses smoke-only dispatch when full-suite fails, and only
marks latest:true for a real five-platform topology. CMT logic is split
into focused packages for reviewability; PR/main E2E server counts are
unchanged.
