+++ server/e2e_dryrun_test.go
@@ -157,20 +157,6 @@ func makeMobileInstances() []*E2EInstance {
 	}
 }
 
-func makeMobileCMTInstances(versions []string) []*E2EInstance {
-	instances := make([]*E2EInstance, 0, len(versions)*len(mobileE2EPlatforms))
-	for versionIndex, version := range versions {
-		for platformIndex, platform := range mobileE2EPlatforms {
-			instances = append(instances, &E2EInstance{
-				Platform:      platform,
-				URL:           fmt.Sprintf("https://v%d-site%d.example.com", versionIndex, platformIndex+1),
-				ServerVersion: version,
-			})
-		}
-	}
-	return instances
-}
-
 // ------------------------------------------------------------
 // 1. Desktop PR label: E2E/Run → e2e-functional.yml
 // ------------------------------------------------------------
@@ -413,213 +399,7 @@ func TestDryRun_DesktopPushEvent(t *testing.T) {
 }
 
 // ------------------------------------------------------------
-// 7. Desktop CMT logic
-// ------------------------------------------------------------
-
-func TestDryRun_DesktopCMT(t *testing.T) {
-	t.Run("parses server_versions input", func(t *testing.T) {
-		versions := parseServerVersionsFromString("v11.1.0, v11.2.0, v12.0.0")
-		assert.Equal(t, []string{"v11.1.0", "v11.2.0", "v12.0.0"}, versions)
-	})
-
-	t.Run("buildDesktopCMTMatrixJSON produces correct schema", func(t *testing.T) {
-		versions := []string{"v11.1.0", "v11.2.0"}
-		instances := []*E2EInstance{
-			{URL: "https://v1.example.com", ServerVersion: "v11.1.0"},
-			{URL: "https://v2.example.com", ServerVersion: "v11.2.0"},
-		}
-		jsonStr, err := buildDesktopCMTMatrixJSON(versions, instances)
-		require.NoError(t, err)
-
-		var matrix map[string]interface{}
-		require.NoError(t, json.Unmarshal([]byte(jsonStr), &matrix))
-
-		environments, ok := matrix["environment"].([]interface{})
-		require.True(t, ok)
-		assert.Len(t, environments, 3, "must have linux, macos, windows")
-		runners := map[string]string{}
-		for _, env := range environments {
-			e := env.(map[string]interface{})
-			runners[e["os"].(string)] = e["runner"].(string)
-		}
-		assert.Equal(t, "ubuntu-latest", runners["linux"])
-		assert.Equal(t, "macos-26", runners["macos"])
-		assert.Equal(t, "windows-2022", runners["windows"])
-
-		servers, ok := matrix["server"].([]interface{})
-		require.True(t, ok)
-		require.Len(t, servers, 2)
-		s0 := servers[0].(map[string]interface{})
-		assert.Equal(t, "v11.1.0", s0["version"])
-		assert.Equal(t, "https://v1.example.com", s0["url"])
-		// Desktop ignores the `latest` field; cmtServer.Latest is shared with mobile but is
-		// never set on the desktop path, and `omitempty` keeps it out of the JSON entirely.
-		_, has0 := s0["latest"]
-		assert.False(t, has0, "desktop matrix must not carry the `latest` field")
-		s1 := servers[1].(map[string]interface{})
-		assert.Equal(t, "v11.2.0", s1["version"])
-		assert.Equal(t, "https://v2.example.com", s1["url"])
-		_, has1 := s1["latest"]
-		assert.False(t, has1, "desktop matrix must not carry the `latest` field")
-	})
-
-	t.Run("CMT tracking key is keyed by dispatched test run id", func(t *testing.T) {
-		repoName := "mattermost-desktop"
-		var testRunID int64 = 999
-		key := cmtInstanceKey(repoName, testRunID)
-		assert.Equal(t, "mattermost-desktop-cmt-999", key)
-	})
-
-	t.Run("CMT workflow name detection", func(t *testing.T) {
-		isCMT := func(name string) bool {
-			return strings.Contains(name, "cmt") || strings.Contains(name, "CMT")
-		}
-		assert.True(t, isCMT("CMT Provisioner"))
-		assert.True(t, isCMT("CMT Mobile"))
-		assert.True(t, isCMT("cmt-workflow"))
-		// The actual test workflow must NOT match — its completion is handled via
-		// isE2ETestWorkflow ("Compatibility Matrix Testing" in E2ETestWorkflowNames)
-		assert.False(t, isCMT("Compatibility Matrix Testing"))
-		assert.False(t, isCMT("E2E Desktop"))
-	})
-}
-
-// ------------------------------------------------------------
-// 8. Mobile CMT logic
-// ------------------------------------------------------------
-
-func TestDryRun_MobileCMT(t *testing.T) {
-	t.Run("buildMobileCMTMatrixJSON produces correct schema", func(t *testing.T) {
-		versions := []string{"v11.1.0", "v11.2.0"}
-		instances := makeMobileCMTInstances(versions)
-		jsonStr, err := buildMobileCMTMatrixJSON(versions, instances)
-		require.NoError(t, err)
-
-		var matrix map[string]interface{}
-		require.NoError(t, json.Unmarshal([]byte(jsonStr), &matrix))
-
-		servers, ok := matrix["server"].([]interface{})
-		require.True(t, ok)
-		require.Len(t, servers, 2)
-		s0 := servers[0].(map[string]interface{})
-		assert.Equal(t, "v11.1.0", s0["version"])
-		assert.Equal(t, "https://v0-site1.example.com", s0["android_site_1_url"])
-		assert.Equal(t, "https://v0-site2.example.com", s0["android_site_2_url"])
-		assert.Equal(t, "https://v0-site3.example.com", s0["ios_site_1_url"])
-		assert.Equal(t, "https://v0-site4.example.com", s0["ios_site_2_url"])
-		assert.Equal(t, "https://v0-site5.example.com", s0["site_3_url"])
-		// `url` stays populated (site-1) for release branches cut before mobile's
-		// five-server CMT rewrite: those workflows read ${{ matrix.server.url }} and would
-		// otherwise test against an empty server URL for a whole release cycle.
-		assert.Equal(t, "https://v0-site3.example.com", s0["url"])
-		// Older version: `latest` is omitted entirely (cmtServer.Latest is false, omitempty).
-		_, has0 := s0["latest"]
-		assert.False(t, has0, "older mobile entries must not carry the `latest` field")
-		s1 := servers[1].(map[string]interface{})
-		assert.Equal(t, "v11.2.0", s1["version"])
-		assert.Equal(t, "https://v1-site1.example.com", s1["android_site_1_url"])
-		assert.Equal(t, "https://v1-site5.example.com", s1["site_3_url"])
-		// Highest semver gets `latest: true`. The mobile workflow uses this to decide whether
-		// to run the whole suite (latest) or just smoke (older) — that policy lives there,
-		// not in matterwick.
-		assert.Equal(t, true, s1["latest"])
-	})
-
-	t.Run("CMT_MATRIX carries explicit five-server topology per version", func(t *testing.T) {
-		versions := []string{"v11.1.0", "v11.2.0"}
-		instances := makeMobileCMTInstances(versions)
-		jsonStr, err := buildMobileCMTMatrixJSON(versions, instances)
-		require.NoError(t, err)
-
-		assert.Contains(t, jsonStr, "\"server\"")
-		assert.Contains(t, jsonStr, "\"android_site_1_url\"")
-		assert.Contains(t, jsonStr, "\"android_site_2_url\"")
-		assert.Contains(t, jsonStr, "\"ios_site_1_url\"")
-		assert.Contains(t, jsonStr, "\"ios_site_2_url\"")
-		assert.Contains(t, jsonStr, "\"site_3_url\"")
-	})
-
-	t.Run("mobile CMT rejects partial topologies", func(t *testing.T) {
-		versions := []string{"v11.1.0", "v11.2.0"}
-		_, err := buildMobileCMTMatrixJSON(versions, makeMobileCMTInstances(versions)[:9])
-		require.EqualError(t, err, "mobile CMT requires 10 instances for 2 versions, got 9")
-	})
-
-	t.Run("mobile CMT marks the highest-semver entry as latest", func(t *testing.T) {
-		// 5-element resolved set (today's typical shape): ESR + 3 minors + current RC.
-		// Across the boundary cases that matter: ESR is older despite high patch numbers;
-		// RC vs stable for the same X.Y.Z should treat stable as higher; multi-digit RC
-		// numbers (rc.10 > rc.2). Locking these in so the workflow's latest gate doesn't
-		// silently shift if someone tweaks the comparator.
-		cases := []struct {
-			name          string
-			versions      []string
-			wantLatestIdx int
-		}{
-			{
-				name:          "ESR + 3 minors + RC: RC's base is newest so RC is latest",
-				versions:      []string{"10.11.19", "11.5.7", "11.6.4", "11.7.2", "11.8.0-rc3"},
-				wantLatestIdx: 4,
-			},
-			{
-				name:          "ESR with high patch loses to lower-patch newer minor",
-				versions:      []string{"10.11.19", "11.0.0"},
-				wantLatestIdx: 1,
-			},
-			{
-				name:          "stable beats same-X.Y.Z RC",
-				versions:      []string{"11.7.0-rc3", "11.7.0"},
-				wantLatestIdx: 1,
-			},
-			{
-				name:          "rc.10 > rc.2 (no string compare)",
-				versions:      []string{"11.8.0-rc2", "11.8.0-rc10"},
-				wantLatestIdx: 1,
-			},
-			{
-				name:          "single version is latest",
-				versions:      []string{"11.7.2"},
-				wantLatestIdx: 0,
-			},
-		}
-		for _, tc := range cases {
-			t.Run(tc.name, func(t *testing.T) {
-				instances := makeMobileCMTInstances(tc.versions)
-				jsonStr, err := buildMobileCMTMatrixJSON(tc.versions, instances)
-				require.NoError(t, err)
-				var matrix map[string]interface{}
-				require.NoError(t, json.Unmarshal([]byte(jsonStr), &matrix))
-				servers := matrix["server"].([]interface{})
-				require.Len(t, servers, len(tc.versions))
-				for i, raw := range servers {
-					s := raw.(map[string]interface{})
-					_, has := s["latest"]
-					if i == tc.wantLatestIdx {
-						assert.Equal(t, true, s["latest"], "index %d (%q) should be latest", i, tc.versions[i])
-					} else {
-						assert.False(t, has, "index %d (%q) must not carry latest", i, tc.versions[i])
-					}
-				}
-			})
-		}
-	})
-
-	t.Run("mobile CMT: all unparseable versions => last entry marked latest", func(t *testing.T) {
-		versions := []string{"junk", "also-junk"}
-		instances := makeMobileCMTInstances(versions)
-		jsonStr, err := buildMobileCMTMatrixJSON(versions, instances)
-		require.NoError(t, err)
-		var matrix map[string]interface{}
-		require.NoError(t, json.Unmarshal([]byte(jsonStr), &matrix))
-		servers := matrix["server"].([]interface{})
-		_, has0 := servers[0].(map[string]interface{})["latest"]
-		assert.False(t, has0)
-		assert.Equal(t, true, servers[1].(map[string]interface{})["latest"])
-	})
-}
-
-// ------------------------------------------------------------
-// 9. Instance tracking and cleanup
+// Instance tracking and cleanup
 // ------------------------------------------------------------
 
 func TestDryRun_InstanceTracking(t *testing.T) {
@@ -1313,180 +1093,6 @@ func TestDryRun_MMServerVersionFromInstance(t *testing.T) {
 	})
 }
 
-// ------------------------------------------------------------
-// 14. CMT version normalization (v-prefix stripping)
-// ------------------------------------------------------------
-
-func TestDryRun_CMTVersionNormalization(t *testing.T) {
-	t.Run("comma-separated input parsed and v-stripped", func(t *testing.T) {
-		// parseServerVersionsFromString splits; the CMT loop then strips v from each.
-		raw := "v11.0.1, v11.1.0, 11.2.0"
-		parsed := parseServerVersionsFromString(raw)
-		require.Len(t, parsed, 3)
-
-		var stripped []string
-		for _, v := range parsed {
-			stripped = append(stripped, strings.TrimPrefix(strings.TrimSpace(v), "v"))
-		}
-		assert.Equal(t, []string{"11.0.1", "11.1.0", "11.2.0"}, stripped)
-	})
-
-	t.Run("CMT matrix JSON contains stripped versions", func(t *testing.T) {
-		// buildDesktopCMTMatrixJSON uses instance.ServerVersion directly.
-		// With stripped versions, the matrix has Docker Hub compatible version strings.
-		versions := []string{"11.0.1", "11.1.0"} // already stripped
-		instances := []*E2EInstance{
-			{URL: "https://11-0-1.example.com", ServerVersion: "11.0.1"},
-			{URL: "https://11-1-0.example.com", ServerVersion: "11.1.0"},
-		}
-		jsonStr, err := buildDesktopCMTMatrixJSON(versions, instances)
-		require.NoError(t, err)
-
-		var matrix map[string]interface{}
-		require.NoError(t, json.Unmarshal([]byte(jsonStr), &matrix))
-
-		servers, ok := matrix["server"].([]interface{})
-		require.True(t, ok)
-		require.Len(t, servers, 2)
-
-		for _, srv := range servers {
-			s := srv.(map[string]interface{})
-			ver := s["version"].(string)
-			assert.False(t, strings.HasPrefix(ver, "v"),
-				"CMT matrix version %q must not have 'v' prefix", ver)
-		}
-
-		s0 := servers[0].(map[string]interface{})
-		s1 := servers[1].(map[string]interface{})
-		assert.Equal(t, "11.0.1", s0["version"])
-		assert.Equal(t, "11.1.0", s1["version"])
-	})
-
-}
-
-// ------------------------------------------------------------
-// 15. resolveCMTServerVersions() — auto-derived CMT version set
-// ------------------------------------------------------------
-
-func TestDryRun_ResolveCMTServerVersions(t *testing.T) {
-	// A realistic releases payload (newest first): an upcoming RC, recent stable minors,
-	// and ESR lines flagged in the body. Includes multiple patches per line, a draft, and
-	// an EOL ESR (9.11) that must not enter the matrix.
-	releasesBody := `[
-		{"tag_name":"v11.8.0-rc3","draft":false,"prerelease":true,"body":"Mattermost Platform Release 11.8.0-rc3"},
-		{"tag_name":"v11.8.0-rc2","draft":false,"prerelease":true,"body":"rc"},
-		{"tag_name":"v11.7.2","draft":false,"prerelease":false,"body":"Mattermost Platform Extended Support Release 11.7.2 contains fixes."},
-		{"tag_name":"v11.7.0","draft":false,"prerelease":false,"body":"Mattermost Platform Extended Support Release 11.7.0"},
-		{"tag_name":"v11.6.4","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.6.4"},
-		{"tag_name":"v11.6.3","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.6.3"},
-		{"tag_name":"v11.5.7","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.5.7"},
-		{"tag_name":"v11.99.0","draft":true,"prerelease":false,"body":"draft should be ignored"},
-		{"tag_name":"v10.11.19","draft":false,"prerelease":false,"body":"Mattermost Platform Extended Support Release 10.11.19 contains security fixes."},
-		{"tag_name":"v10.11.17","draft":false,"prerelease":false,"body":"Mattermost Platform Extended Support Release 10.11.17"},
-		{"tag_name":"v9.11.18","draft":false,"prerelease":false,"body":"Mattermost Platform Extended Support Release 9.11.18"}
-	]`
-
-	t.Run("auto-derives newest 2 ESRs + latest 3 minors + current RC, latest patch each", func(t *testing.T) {
-		srv := mockReleasesServer(t, releasesBody, http.StatusOK)
-		s := newDryRunServer(t, "", "mattermost")
-		s.githubAPIBase = srv.URL + "/"
-
-		got := s.resolveCMTServerVersions()
-		// 10.11.19 + 11.7.2 (newest 2 ESRs) + 11.5.7/11.6.4 (fill latest-3) + 11.8.0-rc3,
-		// 9.11.18 EOL ESR dropped, latest patch per line, v-stripped, ascending.
-		assert.Equal(t, []string{"10.11.19", "11.5.7", "11.6.4", "11.7.2", "11.8.0-rc3"}, got)
-	})
-
-	t.Run("cmtServerVersions uses resolve when CMTServerVersions is empty", func(t *testing.T) {
-		srv := mockReleasesServer(t, releasesBody, http.StatusOK)
-		s := newDryRunServer(t, "", "mattermost")
-		s.githubAPIBase = srv.URL + "/"
-		s.Config.CMTServerVersions = nil
-
-		assert.Equal(t, []string{"10.11.19", "11.5.7", "11.6.4", "11.7.2", "11.8.0-rc3"}, s.cmtServerVersions("desktop"))
-	})
-
-	t.Run("explicit CMTServerVersions override skips resolve", func(t *testing.T) {
-		called := false
-		srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			called = true
-			_, _ = w.Write([]byte(releasesBody))
-		}))
-		t.Cleanup(srv.Close)
-		s := newDryRunServer(t, "", "mattermost")
-		s.githubAPIBase = srv.URL + "/"
-		s.Config.CMTServerVersions = []string{"10.11.22", "11.10.0-rc1"}
-
-		assert.Equal(t, []string{"10.11.22", "11.10.0-rc1"}, s.cmtServerVersions("desktop"))
-		assert.Equal(t, []string{"10.11.22", "11.10.0-rc1"}, s.cmtServerVersions("mobile"), "manual override applies to mobile too")
-		assert.False(t, called, "manual override must not hit the GitHub API")
-	})
-
-	t.Run("API error falls back to defaultCMTServerVersions", func(t *testing.T) {
-		srv := mockReleasesServer(t, "boom", http.StatusInternalServerError)
-		s := newDryRunServer(t, "", "mattermost")
-		s.githubAPIBase = srv.URL + "/"
-		s.Config.CMTServerVersions = nil
-
-		assert.Equal(t, defaultCMTServerVersions, s.resolveCMTServerVersions())
-		assert.Equal(t, []string{"10.11.22", "11.7.7"}, defaultCMTServerVersions)
-		assert.Equal(t, defaultCMTServerVersions, s.cmtServerVersions("desktop"))
-		assert.Equal(t, defaultCMTServerVersions, s.cmtServerVersions("mobile"))
-	})
-
-	t.Run("cap prefers trailing ESR over oldest feature minor", func(t *testing.T) {
-		// 2 ESRs + 3 distinct latest minors + RC = 6 before cap; drop 11.8 (oldest non-ESR).
-		body := `[
-			{"tag_name":"v11.11.0-rc1","draft":false,"prerelease":true,"body":"rc"},
-			{"tag_name":"v11.10.0","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.10.0"},
-			{"tag_name":"v11.9.0","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.9.0"},
-			{"tag_name":"v11.8.0","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.8.0"},
-			{"tag_name":"v11.7.7","draft":false,"prerelease":false,"body":"Mattermost Platform Extended Support Release 11.7.7"},
-			{"tag_name":"v10.11.22","draft":false,"prerelease":false,"body":"Mattermost Platform Extended Support Release 10.11.22"}
-		]`
-		srv := mockReleasesServer(t, body, http.StatusOK)
-		s := newDryRunServer(t, "", "mattermost")
-		s.githubAPIBase = srv.URL + "/"
-
-		got := s.resolveCMTServerVersions()
-		assert.Equal(t, []string{"10.11.22", "11.7.7", "11.9.0", "11.10.0", "11.11.0-rc1"}, got)
-		assert.NotContains(t, got, "11.8.0")
-		assert.Len(t, got, maxCMTServerVersions)
-	})
-
-	t.Run("RC omitted when not newer than latest stable", func(t *testing.T) {
-		// Only stable releases here; an old RC for an already-released line must not appear.
-		body := `[
-			{"tag_name":"v11.7.2","draft":false,"prerelease":false,"body":"Mattermost Platform Extended Support Release 11.7.2"},
-			{"tag_name":"v11.7.0-rc1","draft":false,"prerelease":true,"body":"rc"},
-			{"tag_name":"v11.6.4","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.6.4"},
-			{"tag_name":"v11.5.7","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.5.7"}
-		]`
-		srv := mockReleasesServer(t, body, http.StatusOK)
-		s := newDryRunServer(t, "", "mattermost")
-		s.githubAPIBase = srv.URL + "/"
-
-		got := s.resolveCMTServerVersions()
-		assert.Equal(t, []string{"11.5.7", "11.6.4", "11.7.2"}, got, "stale RC must be excluded")
-	})
-
-	t.Run("parseCMTVersion handles stable, rc, and v-prefix; rejects junk", func(t *testing.T) {
-		v, ok := parseCMTVersion("v11.8.0-rc3")
-		assert.True(t, ok)
-		assert.Equal(t, "11.8.0-rc3", v.raw)
-		assert.Equal(t, 3, v.rc)
-		v2, ok2 := parseCMTVersion("10.11.19")
-		assert.True(t, ok2)
-		assert.Equal(t, 0, v2.rc)
-		_, ok3 := parseCMTVersion("v11.7.0-beta.1")
-		assert.False(t, ok3, "non-rc prerelease suffixes are not CMT versions")
-		_, ok4 := parseCMTVersion("not-a-version")
-		assert.False(t, ok4)
-		// stable sorts above its rc for the same X.Y.Z
-		assert.True(t, v.less(v2) == false)
-	})
-}
-
 // TestResolveBranchHeadSHA verifies the dispatch-time HEAD resolution used to key non-PR
 // cleanup. Non-PR flows dispatch the test workflow with ref=branch, so the run's head_sha is
 // the branch HEAD at dispatch time. We key cleanup on that resolved SHA (not the trigger SHA)
@@ -1684,90 +1290,3 @@ func TestIsBuildReleaseBranch(t *testing.T) {
 	}
 }
 
-// TestDryRun_MobileCMTVersionSelection covers the mobile version set: newest ESR, latest
-// production (newest non-ESR stable), and the current RC — one topology of five servers each.
-func TestDryRun_MobileCMTVersionSelection(t *testing.T) {
-	releasesBody := `[
-		{"tag_name":"v11.8.0-rc3","draft":false,"prerelease":true,"body":"rc"},
-		{"tag_name":"v11.8.0-rc2","draft":false,"prerelease":true,"body":"rc"},
-		{"tag_name":"v11.7.2","draft":false,"prerelease":false,"body":"Mattermost Platform Extended Support Release 11.7.2"},
-		{"tag_name":"v11.7.1","draft":false,"prerelease":false,"body":"Mattermost Platform Extended Support Release 11.7.1"},
-		{"tag_name":"v11.6.4","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.6.4"},
-		{"tag_name":"v11.5.7","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.5.7"},
-		{"tag_name":"v10.11.19","draft":false,"prerelease":false,"body":"Mattermost Platform Extended Support Release 10.11.19"}
-	]`
-
-	t.Run("picks newest ESR + latest production + current RC", func(t *testing.T) {
-		srv := mockReleasesServer(t, releasesBody, http.StatusOK)
-		s := newDryRunServer(t, "", "mattermost")
-		s.githubAPIBase = srv.URL + "/"
-
-		// 11.7.2 is the newest ESR line, 11.6.4 the newest non-ESR stable, 11.8.0-rc3 the RC.
-		// 10.11.19 (older ESR) and 11.5.7 (older stable) are left out.
-		assert.Equal(t, []string{"11.6.4", "11.7.2", "11.8.0-rc3"}, s.resolveMobileCMTServerVersions())
-	})
-
-	t.Run("mobile selection is used for mobile and not for desktop", func(t *testing.T) {
-		srv := mockReleasesServer(t, releasesBody, http.StatusOK)
-		s := newDryRunServer(t, "", "mattermost")
-		s.githubAPIBase = srv.URL + "/"
-		s.Config.CMTServerVersions = nil
-
-		mobile := s.cmtServerVersions("mobile")
-		desktop := s.cmtServerVersions("desktop")
-		assert.Len(t, mobile, maxMobileCMTServerVersions)
-		assert.Greater(t, len(desktop), len(mobile), "desktop keeps its wider set")
-	})
-
-	t.Run("no RC in flight backfills with the next stable line", func(t *testing.T) {
-		body := `[
-			{"tag_name":"v11.7.2","draft":false,"prerelease":false,"body":"Mattermost Platform Extended Support Release 11.7.2"},
-			{"tag_name":"v11.6.4","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.6.4"},
-			{"tag_name":"v11.5.7","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.5.7"}
-		]`
-		srv := mockReleasesServer(t, body, http.StatusOK)
-		s := newDryRunServer(t, "", "mattermost")
-		s.githubAPIBase = srv.URL + "/"
-
-		// ESR 11.7.2 + latest production 11.6.4, then 11.5.7 backfills the empty RC slot.
-		assert.Equal(t, []string{"11.5.7", "11.6.4", "11.7.2"}, s.resolveMobileCMTServerVersions())
-	})
-
-	t.Run("an RC older than the newest stable is not selected", func(t *testing.T) {
-		body := `[
-			{"tag_name":"v11.6.0-rc1","draft":false,"prerelease":true,"body":"stale rc"},
-			{"tag_name":"v11.7.2","draft":false,"prerelease":false,"body":"Mattermost Platform Extended Support Release 11.7.2"},
-			{"tag_name":"v11.8.1","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.8.1"}
-		]`
-		srv := mockReleasesServer(t, body, http.StatusOK)
-		s := newDryRunServer(t, "", "mattermost")
-		s.githubAPIBase = srv.URL + "/"
-
-		got := s.resolveMobileCMTServerVersions()
-		assert.NotContains(t, got, "11.6.0-rc1", "a stale RC must not take the RC slot")
-		assert.Contains(t, got, "11.7.2")
-		assert.Contains(t, got, "11.8.1")
-	})
-
-	t.Run("no ESR flagged still fills the budget from stable lines", func(t *testing.T) {
-		body := `[
-			{"tag_name":"v11.8.1","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.8.1"},
-			{"tag_name":"v11.7.2","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.7.2"},
-			{"tag_name":"v11.6.4","draft":false,"prerelease":false,"body":"Mattermost Platform Release 11.6.4"}
-		]`
-		srv := mockReleasesServer(t, body, http.StatusOK)
-		s := newDryRunServer(t, "", "mattermost")
-		s.githubAPIBase = srv.URL + "/"
-
-		assert.Equal(t, []string{"11.6.4", "11.7.2", "11.8.1"}, s.resolveMobileCMTServerVersions())
-	})
-
-	t.Run("API error falls back to the default set", func(t *testing.T) {
-		srv := mockReleasesServer(t, "boom", http.StatusInternalServerError)
-		s := newDryRunServer(t, "", "mattermost")
-		s.githubAPIBase = srv.URL + "/"
-
-		assert.Equal(t, defaultCMTServerVersions, s.resolveMobileCMTServerVersions())
-		assert.LessOrEqual(t, len(defaultCMTServerVersions), maxMobileCMTServerVersions)
-	})
-}
