+++ server/e2e_tests.go
@@ -9,8 +9,6 @@ import (
 	"fmt"
 	"net/url"
 	"os"
-	"sort"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -973,363 +971,6 @@ func (s *Server) resolveBranchHeadSHA(owner, repoName, branch string) (string, e
 		return "", fmt.Errorf("empty SHA for %s/%s@%s", owner, repoName, branch)
 	}
 	return commit.SHA, nil
-}
-
-// cmtVersion is a parsed Mattermost release version: major.minor.patch with an optional
-// release-candidate number. raw is the bare-semver string passed to the cloud provisioner
-// (e.g. "10.11.22" or "11.10.0-rc1").
-type cmtVersion struct {
-	major, minor, patch int
-	rc                  int // 0 = stable, >0 = -rcN
-	raw                 string
-}
-
-type cmtMinorKey struct{ major, minor int }
-
-// parseCMTVersion parses "vX.Y.Z" or "vX.Y.Z-rcN" (the leading "v" is optional). It returns
-// ok=false for anything else (other prerelease suffixes like -beta/-alpha are ignored for CMT).
-func parseCMTVersion(tag string) (cmtVersion, bool) {
-	raw := strings.TrimPrefix(strings.TrimSpace(tag), "v")
-	base := raw
-	rc := 0
-	if i := strings.Index(base, "-rc"); i != -1 {
-		n, err := strconv.Atoi(base[i+len("-rc"):])
-		if err != nil {
-			return cmtVersion{}, false
-		}
-		rc = n
-		base = base[:i]
-	} else if strings.Contains(base, "-") {
-		return cmtVersion{}, false
-	}
-	parts := strings.Split(base, ".")
-	if len(parts) != 3 {
-		return cmtVersion{}, false
-	}
-	maj, err1 := strconv.Atoi(parts[0])
-	min, err2 := strconv.Atoi(parts[1])
-	pat, err3 := strconv.Atoi(parts[2])
-	if err1 != nil || err2 != nil || err3 != nil {
-		return cmtVersion{}, false
-	}
-	return cmtVersion{major: maj, minor: min, patch: pat, rc: rc, raw: raw}, true
-}
-
-// less reports whether a sorts before b by (major, minor, patch, rc). For the same X.Y.Z, a
-// stable release (rc==0) sorts above its release candidates (e.g. 11.8.0-rc3 < 11.8.0).
-func (a cmtVersion) less(b cmtVersion) bool {
-	if a.major != b.major {
-		return a.major < b.major
-	}
-	if a.minor != b.minor {
-		return a.minor < b.minor
-	}
-	if a.patch != b.patch {
-		return a.patch < b.patch
-	}
-	ar, br := a.rc, b.rc
-	if ar == 0 {
-		ar = int(^uint(0) >> 1) // treat stable as the highest "rc" for the same patch
-	}
-	if br == 0 {
-		br = int(^uint(0) >> 1)
-	}
-	return ar < br
-}
-
-const maxCMTServerVersions = 5
-
-// maxMobileCMTServerVersions caps the mobile CMT version set at the three categories that
-// matter: newest ESR, latest production, current RC (see resolveMobileCMTServerVersions).
-// Mobile provisions one len(mobileE2EPlatforms)-server topology per version, so each slot
-// costs five installations — 3 × 5 = 15 per run, against #92's 25. Raising this is
-// expensive; prefer changing which categories are selected over adding slots.
-const maxMobileCMTServerVersions = 3
-
-const maxCMTESRLines = 2 // current + trailing ESR; older body-flagged ESRs are treated as EOL
-
-// cmtServerVersions returns the version set CMT runs against. An explicit, non-empty
-// Config.CMTServerVersions is used verbatim (manual override / pin). Otherwise the set is
-// auto-derived from Mattermost GitHub releases. Shared by mobile and desktop CMT triggers.
-func (s *Server) cmtServerVersions(instanceType string) []string {
-	if len(s.Config.CMTServerVersions) > 0 {
-		return s.Config.CMTServerVersions
-	}
-	if instanceType == "mobile" {
-		return s.resolveMobileCMTServerVersions()
-	}
-	return s.resolveCMTServerVersions()
-}
-
-// cmtReleaseSet is the classified Mattermost release data both CMT selectors work from:
-// newest patch per stable minor line, which of those lines are ESR, and the current RC.
-type cmtReleaseSet struct {
-	latestStable map[cmtMinorKey]cmtVersion
-	esrMinors    map[cmtMinorKey]bool
-	bestRC       cmtVersion
-	haveRC       bool
-}
-
-// newestStableMinors returns every stable line's newest patch, newest line first.
-func (rs cmtReleaseSet) newestStableMinors() []cmtVersion {
-	minors := make([]cmtVersion, 0, len(rs.latestStable))
-	for _, v := range rs.latestStable {
-		minors = append(minors, v)
-	}
-	sort.Slice(minors, func(i, j int) bool { return minors[j].less(minors[i]) })
-	return minors
-}
-
-// isESRLine reports whether v's minor line is flagged as an extended support release.
-func (rs cmtReleaseSet) isESRLine(v cmtVersion) bool {
-	return rs.esrMinors[cmtMinorKey{v.major, v.minor}]
-}
-
-// resolveMobileCMTServerVersions picks the three versions mobile CMT runs against: the
-// newest ESR, the newest non-ESR stable release, and the current RC. Mobile pays five
-// servers per version, so the set is chosen by category rather than by "newest N" — the
-// point is to span ESR → GA → upcoming, not to test three adjacent patches. Falls back to
-// defaultCMTServerVersions when releases can't be fetched.
-func (s *Server) resolveMobileCMTServerVersions() []string {
-	releaseSet, err := s.fetchCMTReleaseSet()
-	if err != nil {
-		s.Logger.WithError(err).Warn("[resolveMobileCMTServerVersions] Failed to classify releases; using default CMT versions")
-		return defaultCMTServerVersions
-	}
-
-	minors := releaseSet.newestStableMinors()
-	if len(minors) == 0 {
-		s.Logger.Warn("[resolveMobileCMTServerVersions] No stable releases parsed; using default CMT versions")
-		return defaultCMTServerVersions
-	}
-
-	seen := map[cmtMinorKey]bool{}
-	chosen := make([]cmtVersion, 0, maxMobileCMTServerVersions)
-	add := func(v cmtVersion) {
-		key := cmtMinorKey{v.major, v.minor}
-		if seen[key] {
-			return
-		}
-		seen[key] = true
-		chosen = append(chosen, v)
-	}
-
-	// Newest ESR line.
-	for _, v := range minors {
-		if releaseSet.isESRLine(v) {
-			add(v)
-			break
-		}
-	}
-	// Newest stable line that isn't the ESR we just took — that's "latest production".
-	for _, v := range minors {
-		if !releaseSet.isESRLine(v) {
-			add(v)
-			break
-		}
-	}
-	// Current RC, only when it's ahead of the newest stable (i.e. an upcoming release).
-	if releaseSet.haveRC && minors[0].less(releaseSet.bestRC) {
-		add(releaseSet.bestRC)
-	}
-
-	// If a category was missing (no ESR flagged, no RC in flight), backfill with the next
-	// newest stable lines so the matrix still uses its full budget.
-	for _, v := range minors {
-		if len(chosen) >= maxMobileCMTServerVersions {
-			break
-		}
-		add(v)
-	}
-
-	sort.Slice(chosen, func(i, j int) bool { return chosen[i].less(chosen[j]) }) // ascending
-	if len(chosen) > maxMobileCMTServerVersions {
-		chosen = chosen[len(chosen)-maxMobileCMTServerVersions:]
-	}
-
-	versions := make([]string, 0, len(chosen))
-	for _, v := range chosen {
-		versions = append(versions, v.raw)
-	}
-	s.Logger.WithField("versions", versions).Info("[resolveMobileCMTServerVersions] Auto-derived mobile CMT server version set (ESR + latest stable + RC)")
-	return versions
-}
-
-// resolveCMTServerVersions fetches Mattermost releases and picks: the newest
-// maxCMTESRLines ESR lines (body-string "extended support release") + latest 3 stable
-// minors + current RC, one patch per line. Only used when Config.CMTServerVersions is
-// empty. Falls back to defaultCMTServerVersions on error.
-func (s *Server) resolveCMTServerVersions() []string {
-	releaseSet, err := s.fetchCMTReleaseSet()
-	if err != nil {
-		s.Logger.WithError(err).Warn("[resolveCMTServerVersions] Failed to classify releases; using default CMT versions")
-		return defaultCMTServerVersions
-	}
-
-	latestStable := releaseSet.latestStable
-	esrMinors := releaseSet.esrMinors
-	bestRC := releaseSet.bestRC
-	haveRC := releaseSet.haveRC
-
-	if len(latestStable) == 0 {
-		s.Logger.Warn("[resolveCMTServerVersions] No stable releases parsed; using default CMT versions")
-		return defaultCMTServerVersions
-	}
-
-	// All stable minor lines, sorted descending (newest first).
-	minors := releaseSet.newestStableMinors()
-
-	selected := map[cmtMinorKey]cmtVersion{}
-	for i := 0; i < len(minors) && i < 3; i++ { // latest 3 stable minor lines
-		selected[cmtMinorKey{minors[i].major, minors[i].minor}] = minors[i]
-	}
-	// Keep only the newest maxCMTESRLines ESR minors (current + trailing). Older lines
-	// still carry "extended support release" in historical GitHub release bodies and
-	// would otherwise flood the matrix (9.11 / 10.5 false positives after EOS).
-	esrChosen := make([]cmtVersion, 0, len(esrMinors))
-	for k := range esrMinors {
-		if v, ok := latestStable[k]; ok {
-			esrChosen = append(esrChosen, v)
-		}
-	}
-	sort.Slice(esrChosen, func(i, j int) bool { return esrChosen[j].less(esrChosen[i]) }) // newest first
-	keptESR := map[cmtMinorKey]bool{}
-	for i := 0; i < len(esrChosen) && i < maxCMTESRLines; i++ {
-		k := cmtMinorKey{esrChosen[i].major, esrChosen[i].minor}
-		selected[k] = esrChosen[i]
-		keptESR[k] = true
-	}
-
-	chosen := make([]cmtVersion, 0, len(selected)+1)
-	for _, v := range selected {
-		chosen = append(chosen, v)
-	}
-	// Include the current RC only when it's newer than the newest stable (an upcoming release).
-	if haveRC && minors[0].less(bestRC) {
-		chosen = append(chosen, bestRC)
-	}
-	sort.Slice(chosen, func(i, j int) bool { return chosen[i].less(chosen[j]) }) // ascending
-
-	// Cap at maxCMTServerVersions. Prefer kept ESR lines over older feature minors so the
-	// trailing ESR is not dropped when latest-3 + 2 ESRs + RC exceeds the cap.
-	if len(chosen) > maxCMTServerVersions {
-		chosen = capCMTVersionsPreferringESR(chosen, keptESR, maxCMTServerVersions)
-	}
-
-	versions := make([]string, 0, len(chosen))
-	for _, v := range chosen {
-		versions = append(versions, v.raw)
-	}
-	s.Logger.WithField("versions", versions).Info("[resolveCMTServerVersions] Auto-derived CMT server version set")
-	return versions
-}
-
-// fetchCMTReleaseSet fetches Mattermost releases and classifies them into the newest patch
-// per stable minor line, which lines are ESR, and the current RC. Shared by the desktop and
-// mobile selectors so both see the same view of releases.
-func (s *Server) fetchCMTReleaseSet() (cmtReleaseSet, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	client := newGithubClient(s.Config.GithubAccessToken)
-	// githubAPIBase is only set in tests to redirect to a mock server.
-	if s.githubAPIBase != "" {
-		if baseURL, parseErr := url.Parse(s.githubAPIBase); parseErr == nil {
-			client.BaseURL = baseURL
-		}
-	}
-
-	var releases []struct {
-		TagName    string `json:"tag_name"`
-		Draft      bool   `json:"draft"`
-		Prerelease bool   `json:"prerelease"`
-		Body       string `json:"body"`
-	}
-	const perPage = 100
-	for page := 1; ; page++ {
-		req, err := client.NewRequest("GET", fmt.Sprintf("/repos/mattermost/mattermost/releases?per_page=%d&page=%d", perPage, page), nil)
-		if err != nil {
-			return cmtReleaseSet{}, fmt.Errorf("failed to build releases request: %w", err)
-		}
-		var pageReleases []struct {
-			TagName    string `json:"tag_name"`
-			Draft      bool   `json:"draft"`
-			Prerelease bool   `json:"prerelease"`
-			Body       string `json:"body"`
-		}
-		if _, err = client.Do(ctx, req, &pageReleases); err != nil {
-			return cmtReleaseSet{}, fmt.Errorf("failed to fetch releases: %w", err)
-		}
-		releases = append(releases, pageReleases...)
-		if len(pageReleases) < perPage {
-			break
-		}
-	}
-
-	latestStable := map[cmtMinorKey]cmtVersion{}
-	esrMinors := map[cmtMinorKey]bool{}
-	var bestRC cmtVersion
-	haveRC := false
-
-	for _, r := range releases {
-		if r.Draft {
-			continue
-		}
-		v, ok := parseCMTVersion(r.TagName)
-		if !ok {
-			continue
-		}
-		key := cmtMinorKey{v.major, v.minor}
-		if v.rc > 0 {
-			if !haveRC || bestRC.less(v) {
-				bestRC = v
-				haveRC = true
-			}
-			continue
-		}
-		if cur, exists := latestStable[key]; !exists || cur.less(v) {
-			latestStable[key] = v
-		}
-		if strings.Contains(strings.ToLower(r.Body), "extended support release") {
-			esrMinors[key] = true
-		}
-	}
-
-	return cmtReleaseSet{
-		latestStable: latestStable,
-		esrMinors:    esrMinors,
-		bestRC:       bestRC,
-		haveRC:       haveRC,
-	}, nil
-}
-
-// capCMTVersionsPreferringESR keeps at most maxN versions from an ascending list, dropping
-// oldest non-ESR entries first. If only ESR lines remain and we're still over, drop the
-// oldest ESR. RC counts as non-ESR but is usually newest so survives.
-func capCMTVersionsPreferringESR(chosen []cmtVersion, esrMinors map[cmtMinorKey]bool, maxN int) []cmtVersion {
-	if len(chosen) <= maxN {
-		return chosen
-	}
-
-	isESR := func(v cmtVersion) bool {
-		return v.rc == 0 && esrMinors[cmtMinorKey{v.major, v.minor}]
-	}
-
-	kept := append([]cmtVersion(nil), chosen...)
-	for len(kept) > maxN {
-		dropIdx := -1
-		for i, v := range kept {
-			if !isESR(v) {
-				dropIdx = i
-				break
-			}
-		}
-		if dropIdx < 0 {
-			dropIdx = 0
-		}
-		kept = append(kept[:dropIdx], kept[dropIdx+1:]...)
-	}
-	return kept
 }
 
 // destroyE2EInstances destroys all given E2E instances
