+++ server/workflow_run.go
@@ -4,13 +4,10 @@
 package server
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"io"
-	"net/url"
 	"regexp"
-	"sort"
 	"strings"
 	"sync"
 	"time"
@@ -149,6 +146,9 @@ func cmtInstanceKey(repoName string, testRunID int64) string {
 func (s *Server) cmtDispatchMutex(repoName string) *sync.Mutex {
 	s.cmtDispatchLocksMu.Lock()
 	defer s.cmtDispatchLocksMu.Unlock()
+	if s.cmtDispatchLocks == nil {
+		s.cmtDispatchLocks = make(map[string]*sync.Mutex)
+	}
 	if m, ok := s.cmtDispatchLocks[repoName]; ok {
 		return m
 	}
@@ -280,543 +280,6 @@ var buildReleaseBranchPattern = regexp.MustCompile(`^build-release-\d{3,4}$`)
 // isBuildReleaseBranch reports whether ref is mobile's RC-cut branch (build-release-NNN). Used as a CMT gate separate from isReleaseBranch to avoid triggering on every cherry-pick.
 func isBuildReleaseBranch(ref string) bool {
 	return buildReleaseBranchPattern.MatchString(ref)
-}
-
-// handleCMTTrigger resolves instance type and server versions, then delegates to handleCMTWithServerVersions.
-func (s *Server) handleCMTTrigger(owner, repoName, branch, sha string, runID int64, logger logrus.FieldLogger) {
-	instanceType := "desktop"
-	if strings.Contains(repoName, "mobile") {
-		instanceType = "mobile"
-	} else if !strings.Contains(repoName, "desktop") {
-		logger.Warn("Repository is neither desktop nor mobile, skipping CMT trigger")
-		return
-	}
-
-	versions := s.cmtServerVersions(instanceType)
-	logger.WithFields(logrus.Fields{
-		"instanceType": instanceType,
-		"versions":     versions,
-	}).Info("Provisioning CMT instances for resolved server versions")
-
-	s.handleCMTWithServerVersions(owner, repoName, instanceType, branch, sha, versions, runID, logger)
-}
-
-// cmtVersionCapFor returns the version cap for instanceType. Mobile provisions a full
-// len(mobileE2EPlatforms)-server topology per version, so its cap must be tighter than
-// desktop's one-server-per-version: 5 versions × 5 servers is 25 installations, which in
-// practice never all reach stable inside one CMT window.
-func cmtVersionCapFor(instanceType string) int {
-	if instanceType == "mobile" {
-		return maxMobileCMTServerVersions
-	}
-	return maxCMTServerVersions
-}
-
-// capCMTServerVersions keeps at most limit entries, preferring the newest parseable
-// semvers. Copies the input so Config.CMTServerVersions is not mutated.
-func capCMTServerVersions(serverVersions []string, limit int) []string {
-	if limit < 1 {
-		limit = 1
-	}
-	if len(serverVersions) <= limit {
-		return serverVersions
-	}
-	sorted := append([]string(nil), serverVersions...)
-	sort.Slice(sorted, func(i, j int) bool {
-		vi, oki := parseCMTVersion(sorted[i])
-		vj, okj := parseCMTVersion(sorted[j])
-		if oki != okj {
-			return !oki // unparseable first so the newest tail keeps valid versions
-		}
-		if !oki {
-			return sorted[i] < sorted[j]
-		}
-		return vi.less(vj)
-	})
-	return sorted[len(sorted)-limit:]
-}
-
-// spanCMTServerVersions reduces serverVersions to at most limit entries while keeping both
-// ends of the range. CMT's whole value is the spread between the oldest supported server
-// (ESR) and the newest (RC) — a plain "keep the newest N" cap would collapse a mobile
-// matrix to two adjacent releases, which tests nothing about compatibility. Unparseable
-// entries are dropped rather than allowed to consume a slot.
-func spanCMTServerVersions(serverVersions []string, limit int) []string {
-	if limit < 1 {
-		limit = 1
-	}
-	if len(serverVersions) <= limit {
-		return serverVersions
-	}
-
-	parseable := make([]string, 0, len(serverVersions))
-	for _, v := range serverVersions {
-		if _, ok := parseCMTVersion(v); ok {
-			parseable = append(parseable, v)
-		}
-	}
-	if len(parseable) == 0 {
-		// Nothing to order by — fall back to the newest-N cap over the raw input.
-		return capCMTServerVersions(serverVersions, limit)
-	}
-	if len(parseable) <= limit {
-		return parseable
-	}
-
-	sort.Slice(parseable, func(i, j int) bool {
-		vi, _ := parseCMTVersion(parseable[i])
-		vj, _ := parseCMTVersion(parseable[j])
-		return vi.less(vj)
-	})
-
-	if limit == 1 {
-		return []string{parseable[len(parseable)-1]}
-	}
-
-	// Evenly spaced picks that always include the oldest and newest entry.
-	last := len(parseable) - 1
-	selected := make([]string, 0, limit)
-	seen := make(map[int]bool, limit)
-	for i := 0; i < limit; i++ {
-		idx := (i*last + (limit-1)/2) / (limit - 1)
-		if seen[idx] {
-			continue
-		}
-		seen[idx] = true
-		selected = append(selected, parseable[idx])
-	}
-	return selected
-}
-
-// handleCMTWithServerVersions orchestrates CMT testing and dispatches compatibility-matrix-testing.yml once.
-func (s *Server) handleCMTWithServerVersions(repoOwner, repoName, instanceType, branch, sha string, serverVersions []string, runID int64, logger logrus.FieldLogger) {
-	// Cap per instance type. Auto-resolve enforces the desktop cap with ESR preference,
-	// but mobile needs a tighter one (5 servers per version) and Config.CMTServerVersions
-	// can bypass auto-resolve entirely.
-	versionCap := cmtVersionCapFor(instanceType)
-	if len(serverVersions) > versionCap {
-		if instanceType == "mobile" {
-			// Keep the ends of the range (oldest ESR + newest RC) so a 2-version mobile
-			// matrix still spans real compatibility distance.
-			originalCount := len(serverVersions)
-			serverVersions = spanCMTServerVersions(serverVersions, versionCap)
-			logger.Warnf("Capping mobile server versions from %d to %d (keeping range ends): %s",
-				originalCount, len(serverVersions), strings.Join(serverVersions, ", "))
-		} else {
-			logger.Warnf("Capping server versions from %d to %d (keeping newest)", len(serverVersions), versionCap)
-			serverVersions = capCMTServerVersions(serverVersions, versionCap)
-		}
-	}
-
-	logger = logger.WithFields(logrus.Fields{
-		"serverVersionCount": len(serverVersions),
-		"branch":             branch,
-		"sha":                sha,
-		"run_id":             runID,
-	})
-	logger.Info("Starting CMT with server versions")
-
-	// Best-effort per version: a version whose topology fails to provision is dropped
-	// (loudly — log + Mattermost alert) and CMT still runs for the versions that came up.
-	// Aborting the whole matrix on one failed installation means zero coverage, which is
-	// strictly worse than reduced coverage, and it is why CMT stopped dispatching at all.
-	var allInstances []*E2EInstance
-	var validVersions []string
-	var droppedVersions []string
-
-	for _, version := range serverVersions {
-		version = strings.TrimSpace(version)
-		if version == "" {
-			continue
-		}
-		// Docker Hub tags use bare semver (e.g. "11.6.0"), not "v11.6.0".
-		version = strings.TrimPrefix(version, "v")
-
-		logger.WithField("version", version).Info("Creating CMT instances for server version")
-
-		var versionInstances []*E2EInstance
-		if instanceType == "mobile" {
-			var err error
-			versionInstances, err = s.createMobileCMTInstances(context.Background(), repoName, version, logger)
-			if err != nil {
-				// createMobileCMTInstances already destroyed its own partial topology.
-				logger.WithError(err).Errorf("Failed to create topology for version %s; dropping this version from the CMT matrix", version)
-				droppedVersions = append(droppedVersions, version)
-				continue
-			}
-		} else {
-			instance, err := s.createSingleCMTInstance(context.Background(), repoName, instanceType, version, "", logger)
-			if err != nil {
-				logger.WithError(err).Errorf("Failed to create instance for version %s; dropping this version from the CMT matrix", version)
-				droppedVersions = append(droppedVersions, version)
-				continue
-			}
-			versionInstances = []*E2EInstance{instance}
-		}
-		expectedInstances := 1
-		if instanceType == "mobile" {
-			expectedInstances = len(mobileE2EPlatforms)
-		}
-		if len(versionInstances) != expectedInstances {
-			logger.Errorf("Incomplete CMT topology for version %s (got %d, want %d); dropping this version", version, len(versionInstances), expectedInstances)
-			s.destroyE2EInstances(versionInstances, logger)
-			droppedVersions = append(droppedVersions, version)
-			continue
-		}
-
-		allInstances = append(allInstances, versionInstances...)
-		validVersions = append(validVersions, version)
-	}
-
-	if len(droppedVersions) > 0 {
-		logger.WithFields(logrus.Fields{
-			"dropped_versions": droppedVersions,
-			"kept_versions":    validVersions,
-		}).Error("CMT matrix is running with reduced server-version coverage")
-		s.logErrorToMattermost("CMT on %s/%s (%s): dropped server version(s) %s — provisioning failed. Running with %s.",
-			repoOwner, repoName, branch, strings.Join(droppedVersions, ", "), strings.Join(validVersions, ", "))
-	}
-
-	if len(allInstances) == 0 {
-		logger.Error("No CMT instances created; nothing to dispatch")
-		if len(droppedVersions) > 0 {
-			s.logErrorToMattermost("CMT on %s/%s (%s) did not run: no server version could be provisioned.", repoOwner, repoName, branch)
-		}
-		return
-	}
-
-	logger.WithField("totalInstances", len(allInstances)).Info("CMT instances created, dispatching test workflow")
-
-	// Build CMT_MATRIX JSON and dispatch compatibility-matrix-testing.yml.
-	var cmtMatrixJSON string
-	var buildErr error
-	if instanceType == "desktop" {
-		cmtMatrixJSON, buildErr = buildDesktopCMTMatrixJSON(validVersions, allInstances)
-	} else {
-		cmtMatrixJSON, buildErr = buildMobileCMTMatrixJSON(validVersions, allInstances)
-	}
-	if buildErr != nil {
-		logger.WithError(buildErr).Error("Failed to build CMT_MATRIX JSON")
-		s.destroyE2EInstances(allInstances, logger)
-		return
-	}
-
-	// Serialize per repo: concurrent triggers could race to the same run id without this mutex.
-	dispatchLock := s.cmtDispatchMutex(repoName)
-	dispatchLock.Lock()
-	defer dispatchLock.Unlock()
-
-	testRunID, err := s.dispatchCMTWorkflow(repoOwner, repoName, branch, cmtMatrixJSON, instanceType, logger)
-	if err != nil {
-		logger.WithError(err).Error("Failed to dispatch compatibility-matrix-testing.yml")
-		s.destroyE2EInstances(allInstances, logger)
-		return
-	}
-
-	if testRunID == 0 {
-		// Dispatch succeeded but run id unresolved — leave instances for the periodic stale-scan.
-		logger.WithField("trigger_run", runID).Warn("CMT dispatched but test run id unresolved; instances left to periodic stale-scan backstop")
-		return
-	}
-
-	key := cmtInstanceKey(repoName, testRunID)
-	s.e2eInstancesLock.Lock()
-	s.e2eInstances[key] = allInstances
-	s.e2eInstancesLock.Unlock()
-
-	logger.WithFields(logrus.Fields{
-		"tracking_key": key,
-		"test_run_id":  testRunID,
-		"trigger_run":  runID,
-	}).Info("CMT workflow dispatched successfully; instances tracked for cleanup")
-}
-
-// createSingleCMTInstance creates one Mattermost cloud instance for a CMT server version.
-func (s *Server) createSingleCMTInstance(ctx context.Context, repoName, instanceType, version, platform string, logger logrus.FieldLogger) (*E2EInstance, error) {
-	sanitizedVersion := sanitizeForDNS(version)
-	// Fresh uid per attempt: a retry must not reuse the failed attempt's DNS name.
-	nameFn := func() string {
-		nameParts := []string{instanceType, sanitizedVersion}
-		if platform != "" {
-			nameParts = append(nameParts, platform)
-		}
-		nameParts = append(nameParts, e2eUniqueSuffix())
-		return e2eInstanceName(s.Config.DNSNameTestServer, nameParts...)
-	}
-
-	username := s.Config.E2EUsername
-	password := s.getE2EPassword(instanceType)
-
-	instance, err := s.createCloudInstallationWithRetry(ctx, nameFn, version, username, password, instanceType, logger)
-	if instance != nil {
-		instance.Platform = platform
-	}
-	return instance, err
-}
-
-// createMobileCMTInstances creates one five-server topology for a server version.
-func (s *Server) createMobileCMTInstances(ctx context.Context, repoName, version string, logger logrus.FieldLogger) ([]*E2EInstance, error) {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	type result struct {
-		instance *E2EInstance
-		err      error
-	}
-	results := make([]result, len(mobileE2EPlatforms))
-	var wg sync.WaitGroup
-	var firstErr error
-	var errMu sync.Mutex
-	for i, platform := range mobileE2EPlatforms {
-		wg.Add(1)
-		go func(idx int, platform string) {
-			defer wg.Done()
-			instance, err := s.createSingleCMTInstance(ctx, repoName, "mobile", version, platform, logger)
-			results[idx] = result{instance: instance, err: err}
-			if err != nil {
-				errMu.Lock()
-				if firstErr == nil {
-					firstErr = err
-					cancel()
-				}
-				errMu.Unlock()
-			}
-		}(i, platform)
-	}
-	wg.Wait()
-
-	instances := make([]*E2EInstance, 0, len(results))
-	for _, result := range results {
-		if result.err != nil {
-			continue
-		}
-		instances = append(instances, result.instance)
-	}
-	if firstErr != nil {
-		s.destroyE2EInstances(instances, logger)
-		return nil, firstErr
-	}
-	return instances, nil
-}
-
-// cmtServer is one entry in CMT_MATRIX. Mobile entries carry a five-server topology.
-type cmtServer struct {
-	Version         string `json:"version"`
-	URL             string `json:"url,omitempty"`
-	AndroidSite1URL string `json:"android_site_1_url,omitempty"`
-	AndroidSite2URL string `json:"android_site_2_url,omitempty"`
-	IOSSite1URL     string `json:"ios_site_1_url,omitempty"`
-	IOSSite2URL     string `json:"ios_site_2_url,omitempty"`
-	Site3URL        string `json:"site_3_url,omitempty"`
-	Latest          bool   `json:"latest,omitempty"`
-}
-
-// buildDesktopCMTMatrixJSON builds CMT_MATRIX for compatibility-matrix-testing.yml: 3 fixed environment runners × N server versions.
-func buildDesktopCMTMatrixJSON(versions []string, instances []*E2EInstance) (string, error) {
-	type cmtEnvironment struct {
-		OS     string `json:"os"`
-		Runner string `json:"runner"`
-	}
-	type desktopCMTMatrix struct {
-		Environment []cmtEnvironment `json:"environment"`
-		Server      []cmtServer      `json:"server"`
-	}
-
-	// macos-13 was retired by GitHub (queues forever). macos-26 matches desktop
-	// PR E2E / CMT remapping. ubuntu-latest matches 24.04 (libasound2t64).
-	matrix := desktopCMTMatrix{
-		Environment: []cmtEnvironment{
-			{OS: "linux", Runner: "ubuntu-latest"},
-			{OS: "macos", Runner: "macos-26"},
-			{OS: "windows", Runner: "windows-2022"},
-		},
-	}
-	for i, version := range versions {
-		if i >= len(instances) {
-			break
-		}
-		matrix.Server = append(matrix.Server, cmtServer{Version: version, URL: instances[i].URL})
-	}
-
-	b, err := json.Marshal(matrix)
-	if err != nil {
-		return "", fmt.Errorf("failed to marshal desktop CMT matrix: %w", err)
-	}
-	return string(b), nil
-}
-
-// buildMobileCMTMatrixJSON builds one five-server topology per version, with the highest semver marked latest:true.
-func buildMobileCMTMatrixJSON(versions []string, instances []*E2EInstance) (string, error) {
-	type mobileCMTMatrix struct {
-		Server []cmtServer `json:"server"`
-	}
-
-	// Mark the highest-parseable version as latest; fall back to last entry if none parse.
-	latestIdx := -1
-	var latestVer cmtVersion
-	requiredInstances := len(versions) * len(mobileE2EPlatforms)
-	if len(instances) != requiredInstances {
-		return "", fmt.Errorf("mobile CMT requires %d instances for %d versions, got %d", requiredInstances, len(versions), len(instances))
-	}
-
-	for i, version := range versions {
-		v, ok := parseCMTVersion(version)
-		if !ok {
-			continue
-		}
-		if latestIdx == -1 || latestVer.less(v) {
-			latestVer = v
-			latestIdx = i
-		}
-	}
-	if latestIdx == -1 && len(versions) > 0 {
-		latestIdx = len(versions) - 1
-	}
-
-	var matrix mobileCMTMatrix
-	for i, version := range versions {
-		entry := cmtServer{Version: version}
-		block := instances[i*len(mobileE2EPlatforms) : (i+1)*len(mobileE2EPlatforms)]
-		platformToURL := make(map[string]string, len(block))
-		for _, inst := range block {
-			platformToURL[inst.Platform] = inst.URL
-		}
-		for _, platform := range mobileE2EPlatforms {
-			url, ok := platformToURL[platform]
-			if !ok {
-				return "", fmt.Errorf("mobile CMT missing instance for platform %s in version %s", platform, version)
-			}
-			switch platform {
-			case "android-site-1":
-				entry.AndroidSite1URL = url
-			case "android-site-2":
-				entry.AndroidSite2URL = url
-			case "ios-site-1":
-				entry.IOSSite1URL = url
-			case "ios-site-2":
-				entry.IOSSite2URL = url
-			case "site-3":
-				entry.Site3URL = url
-			}
-		}
-		// Back-compat: release branches cut before mobile's five-server CMT rewrite read
-		// ${{ matrix.server.url }}. Those branches are live for a whole release cycle, so
-		// keep url populated (site-1) or CMT on them runs against an empty server URL.
-		entry.URL = entry.IOSSite1URL
-		if i == latestIdx {
-			entry.Latest = true
-		}
-		matrix.Server = append(matrix.Server, entry)
-	}
-
-	b, err := json.Marshal(matrix)
-	if err != nil {
-		return "", fmt.Errorf("failed to marshal mobile CMT matrix: %w", err)
-	}
-	return string(b), nil
-}
-
-// dispatchCMTWorkflow dispatches compatibility-matrix-testing.yml and polls for the run id used to key cleanup.
-func (s *Server) dispatchCMTWorkflow(repoOwner, repoName, branch, cmtMatrixJSON, instanceType string, logger logrus.FieldLogger) (int64, error) {
-	ctx := context.Background()
-	client := newGithubClient(s.Config.GithubAccessToken)
-
-	// Snapshot run ids before dispatch so the poll only accepts strictly-new ids.
-	preDispatchRunIDs, snapErr := s.listExistingCMTRunIDs(repoOwner, repoName, branch, logger)
-	if snapErr != nil {
-		// Non-fatal: poll will still accept any new id that appears after this moment.
-		logger.WithError(snapErr).Warn("Failed to snapshot pre-dispatch CMT run ids; poll will only accept ids not seen before dispatch")
-		preDispatchRunIDs = map[int64]bool{}
-	}
-
-	workflowInputs := map[string]interface{}{
-		"CMT_MATRIX": cmtMatrixJSON,
-	}
-	if instanceType == "desktop" {
-		workflowInputs["DESKTOP_VERSION"] = branch
-	} else {
-		workflowInputs["MOBILE_VERSION"] = branch
-	}
-
-	logger.WithFields(logrus.Fields{
-		"ref":          branch,
-		"instanceType": instanceType,
-	}).Debug("Dispatching compatibility-matrix-testing.yml")
-
-	req, err := client.NewRequest("POST",
-		fmt.Sprintf("/repos/%s/%s/actions/workflows/compatibility-matrix-testing.yml/dispatches", repoOwner, repoName),
-		map[string]interface{}{
-			"ref":    branch,
-			"inputs": workflowInputs,
-		})
-	if err != nil {
-		return 0, fmt.Errorf("failed to create CMT workflow dispatch request: %w", err)
-	}
-
-	resp, err := client.Do(ctx, req, nil)
-	if err != nil {
-		return 0, fmt.Errorf("failed to dispatch compatibility-matrix-testing.yml: %w", err)
-	}
-	if resp.StatusCode != 204 {
-		return 0, fmt.Errorf("unexpected status %d from compatibility-matrix-testing.yml dispatch", resp.StatusCode)
-	}
-
-	testRunID, err := s.pollDispatchedWorkflowRun(repoOwner, repoName, "compatibility-matrix-testing.yml", branch, preDispatchRunIDs, logger)
-	if err != nil {
-		// Return 0 to skip run-id tracking; leave instances for the periodic stale-scan.
-		logger.WithError(err).Warn("Dispatched compatibility-matrix-testing.yml but could not resolve test run id within poll deadline")
-		return 0, nil
-	}
-
-	logger.WithField("test_run_id", testRunID).Info("compatibility-matrix-testing.yml dispatched successfully")
-	return testRunID, nil
-}
-
-// listExistingCMTRunIDs returns recent run ids for compatibility-matrix-testing.yml on branch, snapshotted before dispatch.
-func (s *Server) listExistingCMTRunIDs(repoOwner, repoName, branch string, logger logrus.FieldLogger) (map[int64]bool, error) {
-	runs, err := s.listCMTRuns(repoOwner, repoName, "compatibility-matrix-testing.yml", branch)
-	if err != nil {
-		return nil, err
-	}
-	ids := make(map[int64]bool, len(runs))
-	for _, run := range runs {
-		ids[run.ID] = true
-	}
-	logger.WithField("pre_dispatch_run_count", len(ids)).Debug("Snapshotted pre-dispatch CMT run ids")
-	return ids, nil
-}
-
-// cmtWorkflowRun is the minimal slice of the GitHub workflow_run object we need.
-type cmtWorkflowRun struct {
-	ID        int64  `json:"id"`
-	CreatedAt string `json:"created_at"`
-}
-
-// listCMTRuns fetches up to the 10 most-recent workflow_dispatch runs for workflowFile on branch.
-func (s *Server) listCMTRuns(repoOwner, repoName, workflowFile, branch string) ([]cmtWorkflowRun, error) {
-	client := newGithubClient(s.Config.GithubAccessToken)
-	if s.githubAPIBase != "" {
-		if baseURL, parseErr := url.Parse(s.githubAPIBase); parseErr == nil {
-			client.BaseURL = baseURL
-		}
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
-	defer cancel()
-
-	listURL := fmt.Sprintf("/repos/%s/%s/actions/workflows/%s/runs?event=workflow_dispatch&branch=%s&per_page=10",
-		repoOwner, repoName, workflowFile, url.QueryEscape(branch))
-	req, err := client.NewRequest("GET", listURL, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create workflow runs list request: %w", err)
-	}
-	var resp struct {
-		WorkflowRuns []cmtWorkflowRun `json:"workflow_runs"`
-	}
-	if _, err = client.Do(ctx, req, &resp); err != nil {
-		return nil, fmt.Errorf("failed to list workflow runs: %w", err)
-	}
-	return resp.WorkflowRuns, nil
 }
 
 // pollDispatchedWorkflowRun polls for a new run id not in preDispatchRunIDs. The pre-dispatch snapshot makes this race-free vs. time-window approaches.
